refactor(sql): move PivotQuery into ast/pivot
Move pivot-specific AST types, helpers, and builder methods out of
ast/query and into a dedicated ast/pivot module.

Keep the existing Query.pivot and WithClause.pivot entry points intact,
and re-export PivotQuery, PivotSource, and isPivotQuery from ast/index.
