+++ packages/mosaic/sql/src/ast/query.ts
@@ -1,10 +1,11 @@
-import type { FilterExpr, FromExpr, GroupByExpr, MaybeArray, OrderByExpr, PivotInExpr, PivotOnExpr, PivotUsingEntry, PivotUsingExpr, SelectExpr, WithExpr } from '../types.js';
+import type { FilterExpr, FromExpr, GroupByExpr, MaybeArray, OrderByExpr, SelectExpr, WithExpr } from '../types.js';
+import type { PivotQuery, PivotSource } from './pivot.js';
 import type { SampleMethod } from './sample.js';
-import { CREATE_QUERY, CREATE_SCHEMA_QUERY, DESCRIBE_QUERY, PIVOT_QUERY, SELECT_QUERY, SET_OPERATION } from '../constants.js';
-import { asLiteral, asNode, asTableRef, asVerbatim, maybeTableRef } from '../util/ast.js';
+import { CREATE_QUERY, CREATE_SCHEMA_QUERY, DESCRIBE_QUERY, SELECT_QUERY, SET_OPERATION } from '../constants.js';
+import { asNode, asTableRef, asVerbatim, maybeTableRef } from '../util/ast.js';
 import { exprList, nodeList } from '../util/function.js';
 import { unquote } from '../util/string.js';
-import { isArray, isParamLike, isString } from '../util/type-check.js';
+import { isArray, isString } from '../util/type-check.js';
 import { isColumnRef } from './column-ref.js';
 import { FromClauseNode, FromNode } from './from.js';
 import { ExprNode, SQLNode, isNode } from './node.js';
@@ -14,6 +15,15 @@ import { isTableRef, TableRefNode } from './table-ref.js';
 import { VerbatimNode } from './verbatim.js';
 import { WindowClauseNode, type WindowDefNode } from './window.js';
 import { WithClauseNode } from './with.js';
+
+// PivotQuery lives in pivot.ts to keep this module focused on the shared query
+// API. It registers this factory to avoid a runtime import cycle: PivotQuery
+// extends Query, while Query.pivot needs to construct a PivotQuery.
+let createPivotQuery: ((source: PivotSource) => PivotQuery) | undefined;
+
+export function setPivotQueryFactory(factory: (source: PivotSource) => PivotQuery): void {
+  createPivotQuery = factory;
+}
 
 /**
  * Check if a value is a selection query or set operation.
@@ -29,14 +39,6 @@ export function isQuery(value: unknown): value is Query {
  */
 export function isSelectQuery(value: unknown): value is SelectQuery {
   return value instanceof SelectQuery;
-}
-
-/**
- * Check if a value is a pivot query.
- * @param value The value to check.
- */
-export function isPivotQuery(value: unknown): value is PivotQuery {
-  return value instanceof PivotQuery;
 }
 
 /**
@@ -93,7 +95,10 @@ export class Query extends ExprNode {
    * @param source The source relation to pivot.
    */
   static pivot(source: PivotSource) {
-    return new PivotQuery(source);
+    if (!createPivotQuery) {
+      throw new Error('PivotQuery factory has not been registered.');
+    }
+    return createPivotQuery(source);
   }
 
   /**
@@ -268,140 +273,6 @@ export class Query extends ExprNode {
   offset(value: number | ExprNode): this {
     this._offset = asNode(value);
     return this;
-  }
-}
-
-export type PivotSource = string | string[] | SQLNode;
-
-export class PivotQuery extends Query {
-  /** The relation to pivot. */
-  readonly source: SQLNode;
-  /** The expressions whose values determine pivot output columns. */
-  _on: ExprNode[] = [];
-  /** The literal values that constrain and order pivot output columns. */
-  _in: ExprNode[] = [];
-  /** The aggregate expressions used to populate pivot output cells. */
-  _using: SelectClauseNode[] = [];
-  /** The expressions that define pivot row groups. */
-  _groupby: ExprNode[] = [];
-
-  /**
-   * Instantiate a new pivot query.
-   * @param source The source relation to pivot.
-   */
-  constructor(source: PivotSource) {
-    super(PIVOT_QUERY);
-    this.source = maybeTableRef(source);
-  }
-
-  /**
-   * Add a pointer to the query for which this query is a CTE.
-   * @param query The query for which this query is a CTE.
-   */
-  setCteFor(query: Query | null): void {
-    super.setCteFor(query);
-    if (isQuery(this.source)) {
-      this.source.setCteFor(query);
-    }
-  }
-
-  /**
-   * Return a list of subqueries.
-   */
-  get subqueries(): Query[] {
-    return isQuery(this.source) ? [this.source] : [];
-  }
-
-  /**
-   * Add ON expressions.
-   * @param expr Expressions to add.
-   */
-  on(...expr: PivotOnExpr[]): this {
-    this._on = this._on.concat(nodeList(expr));
-    return this;
-  }
-
-  /**
-   * Add IN values to constrain and order pivot output columns.
-   * @param expr Values to add.
-   */
-  in(...expr: [PivotInExpr, ...PivotInExpr[]]): this {
-    const list = expr.flat().map(asLiteral);
-    if (list.length === 0) {
-      throw new Error('PivotQuery.in requires at least one value.');
-    }
-    this._in = this._in.concat(list);
-    return this;
-  }
-
-  /**
-   * Add USING expressions.
-   * @param expr Aggregate cell expressions to add.
-   */
-  using(...expr: [PivotUsingExpr, ...PivotUsingExpr[]]): this {
-    const list: SelectClauseNode[] = [];
-
-    const add = (v: unknown, as?: string) => {
-      if (v != null) {
-        list.push(new SelectClauseNode(asNode(v), as ? unquote(as)! : ''));
-      }
-    };
-
-    const visit = (e: PivotUsingEntry) => {
-      if (e == null) return;
-      else if (
-        isString(e)
-        || isNode(e)
-        || isParamLike(e)
-        || e instanceof Date
-        || typeof e !== 'object'
-      ) {
-        add(e);
-      } else {
-        for (const alias in e) add(e[alias], alias);
-      }
-    };
-
-    expr.forEach(e => isArray(e) ? e.forEach(visit) : visit(e));
-    if (list.length === 0) {
-      throw new Error('PivotQuery.using requires at least one expression.');
-    }
-    this._using = this._using.concat(list);
-    return this;
-  }
-
-  /**
-   * Add GROUP BY expressions.
-   * @param expr Expressions to add.
-   */
-  groupby(...expr: GroupByExpr[]): this {
-    this._groupby = this._groupby.concat(nodeList(expr));
-    return this;
-  }
-
-  /**
-   * Set GROUP BY expressions, replacing any prior expressions.
-   * @param expr Expressions to add.
-   */
-  setGroupby(...expr: GroupByExpr[]): this {
-    this._groupby = [];
-    return this.groupby(...expr);
-  }
-
-  /**
-   * Clone this pivot query.
-   */
-  clone(): this {
-    const { source, ...rest } = this;
-    // @ts-expect-error creates pivot query
-    return Object.assign(new PivotQuery(source), rest, {
-      _with: this._with.slice(),
-      _on: this._on.slice(),
-      _in: this._in.slice(),
-      _using: this._using.slice(),
-      _groupby: this._groupby.slice(),
-      _orderby: this._orderby.slice()
-    });
   }
 }
